Update fusion filter
Previously, the output star-fusion-gene-list-filtered.final only included fusions with both gene partners on the keep list, and now it includes fusions with either partner on the keep list. Specifically, the filter code excluded fusions when either gene wasn't on the "keep" list; now it only excludes fusions when neither gene is on the "keep" list.